Kernel: Move x86 IO instructions code into the x86 specific folder
