+++ Kernel/Bus/PCI/Access.cpp
@@ -7,10 +7,10 @@
 #include <AK/ByteReader.h>
 #include <AK/HashTable.h>
 #include <Kernel/API/KResult.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Firmware/ACPI/Definitions.h>
-#include <Kernel/IO.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/Region.h>
 #include <Kernel/Sections.h>
