+++ Kernel/Devices/HID/I8042Controller.cpp
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Devices/HID/I8042Controller.h>
 #include <Kernel/Devices/HID/PS2KeyboardDevice.h>
 #include <Kernel/Devices/HID/PS2MouseDevice.h>
 #include <Kernel/Devices/HID/VMWareMouseDevice.h>
-#include <Kernel/IO.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
