+++ Kernel/Devices/HID/PS2MouseDevice.cpp
@@ -5,11 +5,11 @@
  */
 
 #include <AK/Memory.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Devices/DeviceManagement.h>
 #include <Kernel/Devices/HID/PS2MouseDevice.h>
 #include <Kernel/Devices/VMWareBackdoor.h>
-#include <Kernel/IO.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
