+++ Kernel/Devices/SB16.cpp
@@ -6,11 +6,11 @@
 
 #include <AK/Memory.h>
 #include <AK/StringView.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Devices/DeviceManagement.h>
 #include <Kernel/Devices/SB16.h>
-#include <Kernel/IO.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Sections.h>
