+++ Kernel/Devices/SerialDevice.h
@@ -6,8 +6,8 @@
 
 #pragma once
 
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Devices/CharacterDevice.h>
-#include <Kernel/IO.h>
 
 namespace Kernel {
