+++ Kernel/Net/E1000NetworkAdapter.h
@@ -7,9 +7,9 @@
 #pragma once
 
 #include <AK/OwnPtr.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/Device.h>
-#include <Kernel/IO.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Random.h>
