+++ Kernel/Panic.cpp
@@ -5,9 +5,9 @@
  */
 
 #include <AK/Format.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Arch/x86/Processor.h>
 #include <Kernel/CommandLine.h>
-#include <Kernel/IO.h>
 #include <Kernel/KSyms.h>
 #include <Kernel/Panic.h>
 #include <Kernel/Thread.h>
