+++ Kernel/Storage/AHCIPortHandler.h
@@ -7,8 +7,8 @@
 #pragma once
 
 #include <AK/RefPtr.h>
+#include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Devices/Device.h>
-#include <Kernel/IO.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Locking/Mutex.h>
 #include <Kernel/Memory/PhysicalPage.h>
